Move server code to separate file
